qol fix

+++ data.py
@@ -9,13 +9,16 @@ cluster = MongoClient(getenv('MONGO_URI'), tlsCAFile=certifi.where())
 db = cluster["Amazon-Price-Tracker"]
 collection = db["Entries"]
 
+def fetch_data():
+    return collection
+
 def view(userID):    
     for doc in collection.find({"entry.user-info.userID": userID}):
-        yield doc.get("entry").get("ASIN"), doc.get("entry").get("price")
+        yield doc["entry"]["ASIN"], doc["entry"]["price"]          
     
 async def add(asin, email, userID, channelID):
-    asin_list = collection.distinct("entry.ASIN")
-    if asin in asin_list: 
+    exists = collection.find_one({"entry.ASIN" : asin})
+    if exists: 
         collection.update_one( 
             {"entry.ASIN" : asin},
             {"$addToSet": {
@@ -27,7 +30,7 @@ async def add(asin, email, userID, channelID):
                 }
             }
         )
-    elif asin not in asin_list:
+    elif not exists:
         price = await Scraper().scrape('price', asin)
         if price is None:
             return False
@@ -46,18 +49,21 @@ async def add(asin, email, userID, channelID):
         collection.insert_one(entry)
 
 def remove(userID, asin):
-    asin_list = collection.distinct("entry.ASIN")
-    if asin in asin_list: 
+    document = collection.find_one({"entry.ASIN" : asin})
+    if document is None:
+        return False
+    users = document["entry"]["user-info"]
+    user = [user for user in users if user["userID"] == userID]
+    if user:
         collection.update_one(
-            {"entry.ASIN" : asin},
-            {"$pull": {
-                    "entry.user-info": {
-                        "userID": userID
+                {"entry.ASIN" : asin},
+                {"$pull": {
+                        "entry.user-info": {
+                            "userID": userID
+                        }
                     }
                 }
-            }
-        )
-    collection.delete_one({ "entry.user-info.0":{"$exists":False}})
-
-def fetch_data():
-    return collection
+            )
+        collection.delete_one({ "entry.user-info.0":{"$exists":False}})
+        return True
+    return False
